+++ src/vs/editor/browser/widget/hoverWidget/hover.css
