+++ src/vs/editor/browser/widget/markdownRenderer/browser/markdownRenderer.ts
