+++ src/vs/editor/browser/widget/markdownRenderer/browser/renderedMarkdown.css
