+++ src/vs/platform/hover/browser/hover.ts
